Update mdai and target folder structure for [ch1271]
