+++ docker/py37/Dockerfile.template
@@ -8,9 +8,6 @@ ARG MDAI_PATH
 COPY requirements.txt /src/requirements.txt
 RUN pip install -r requirements.txt
 
-COPY lib/requirements.txt /src/model-requirements.txt
-RUN pip install -r model-requirements.txt
-
 COPY lib/${MDAI_PATH}/requirements.txt /src/mdai-requirements.txt
 RUN pip install -r mdai-requirements.txt
